Insertion of performance code in a submodule.
