+++ linera-views/src/test_utils/performance.rs
@@ -1,0 +1,259 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use std::{
+    fmt::Debug,
+    time::{Duration, Instant},
+};
+
+use criterion::black_box;
+use crate::{
+    batch::Batch,
+    test_utils::{add_prefix, get_random_key_values2},
+    store::LocalKeyValueStore,
+};
+
+// We generate about 2000 keys of length 11 with a key of length 10000
+// The keys are of the form 0,x_1, ..., x_n with 0 <= x_i < 4 and n=10.
+
+/// A value to use for the keys
+const PREFIX: &[u8] = &[0];
+
+/// A value to use for the keys
+const PREFIX_SEARCH: &[u8] = &[0, 0];
+
+/// The number of keys
+const NUM_ENTRIES: usize = 200;
+
+/// The number of inserted keys
+const NUM_INSERT: usize = 70;
+
+/// The length of the keys
+const LEN_KEY: usize = 10;
+
+/// The length of the values
+const LEN_VALUE: usize = 10000;
+
+async fn clear_store<S: LocalKeyValueStore>(store: &S) {
+    let mut batch = Batch::new();
+    batch.delete_key_prefix(PREFIX.to_vec());
+    store.write_batch(batch).await.unwrap();
+}
+
+/// Compute benchmark of the `contains_key` operation.
+pub async fn performance_contains_key<S: LocalKeyValueStore>(store: S, iterations: u64) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values[..NUM_INSERT] {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+        store.write_batch(batch).await.unwrap();
+
+        let measurement = Instant::now();
+        for key_value in &key_values {
+            black_box(store.contains_key(&key_value.0).await.unwrap());
+        }
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
+/// Compute benchmark of the `contains_keys` operation.
+pub async fn performance_contains_keys<S: LocalKeyValueStore>(store: S, iterations: u64) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values[..NUM_INSERT] {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+        store.write_batch(batch).await.unwrap();
+	let keys = key_values
+            .into_iter()
+            .map(|(key, _)| key)
+            .collect::<Vec<_>>();
+
+        let measurement = Instant::now();
+        black_box(store.contains_keys(keys).await.unwrap());
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
+/// Compute benchmark of the `performance_find_keys_by_prefix` operation.
+pub async fn performance_find_keys_by_prefix<S: LocalKeyValueStore>(
+    store: S,
+    iterations: u64,
+) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+        store.write_batch(batch).await.unwrap();
+
+        let measurement = Instant::now();
+        black_box(store.find_keys_by_prefix(PREFIX_SEARCH).await.unwrap());
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
+/// Compute benchmark of the `performance_find_keys_by_prefix` operation.
+pub async fn performance_find_key_values_by_prefix<S: LocalKeyValueStore>(
+    store: S,
+    iterations: u64,
+) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+        store.write_batch(batch).await.unwrap();
+
+        let measurement = Instant::now();
+        black_box(
+            store
+                .find_key_values_by_prefix(PREFIX_SEARCH)
+                .await
+                .unwrap(),
+	);
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
+/// Compute benchmark of the `read_value_bytes` operation.
+pub async fn performance_read_value_bytes<S: LocalKeyValueStore>(
+    store: S,
+    iterations: u64,
+) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+        store.write_batch(batch).await.unwrap();
+
+        let measurement = Instant::now();
+        for (key, _) in &key_values {
+            black_box(store.read_value_bytes(key).await.unwrap());
+        }
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
+/// Compute benchmark of the `read_multi_values_bytes` operation.
+pub async fn performance_read_multi_values_bytes<S: LocalKeyValueStore>(
+    store: S,
+    iterations: u64,
+) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+        store.write_batch(batch).await.unwrap();
+        let keys = key_values
+            .into_iter()
+            .map(|(key, _)| key)
+            .collect::<Vec<_>>();
+
+        let measurement = Instant::now();
+        black_box(store.read_multi_values_bytes(keys).await.unwrap());
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
+/// Compute benchmark of the `write_batch` operation.
+pub async fn performance_write_batch<S: LocalKeyValueStore>(store: S, iterations: u64) -> Duration
+where
+    S::Error: Debug,
+{
+    let mut total_time = Duration::ZERO;
+    for _ in 0..iterations {
+        let key_values = add_prefix(
+            PREFIX,
+            get_random_key_values2(NUM_ENTRIES, LEN_KEY, LEN_VALUE),
+        );
+        let mut batch = Batch::new();
+        for key_value in &key_values {
+            batch.put_key_value_bytes(key_value.0.clone(), key_value.1.clone());
+        }
+
+	let measurement = Instant::now();
+        store.write_batch(batch).await.unwrap();
+        total_time += measurement.elapsed();
+
+        clear_store(&store).await;
+    }
+
+    total_time
+}
+
